Add Bind and Connect socket methods.
Connect accepts a System.Uri endpoint, which get translated into a
ZMQ-compatible string. This enforces that the addresses passed to
connect are valid URIs, though it leaves scheme and authority
validation for libzmq.

Bind can only accept a string endpoint, as ZMQ allows addresses formatted as
'tcp://*:4000', which are not valid URIs.